code/server/requirements.txt
certifi==2019.3.9

Flask==2.2.2
Werkzeug==2.2.2
idna==2.8
python-dotenv==0.10.3
requests==2.28.1
toml==0.10.2
urllib3==1.26.13
stripe==4.2.0
Flask-Caching==2.0.1
black==22.12.0
pre-commit==2.21.0
Flask-Cors==5.0.0